atualizando id pra long

+++ EloDrinksAPI/DTOs/item/ItemResponseDto.cs
@@ -7,7 +7,7 @@ namespace EloDrinksAPI.DTOs.item
 {
     public class ItemResponseDto
     {
-        public int IdItem { get; set; }
+        public long IdItem { get; set; }
         public string Nome { get; set; } = null!;
         public string Descricao { get; set; } = null!;
         public float Preco { get; set; }
