+++ EloDrinksAPI/DTOs/orcamentoHasItem/CreateOrcamentoHasItemDto.cs
@@ -2,9 +2,9 @@ namespace EloDrinksAPI.DTOs.orcamentoHasItem
 {
     public class CreateOrcamentoHasItemDto
     {
-        public int OrcamentoIdOrcamento { get; set; }
-        public int OrcamentoUsuarioIdUsuario { get; set; }
-        public int ItemIdItem { get; set; }
+        public long OrcamentoIdOrcamento { get; set; }
+        public long OrcamentoUsuarioIdUsuario { get; set; }
+        public long ItemIdItem { get; set; }
         public int Quantidade { get; set; }
     }
