+++ EloDrinksAPI/DTOs/usuario/UsuarioResponseDto.cs
@@ -2,7 +2,7 @@ namespace EloDrinksAPI.DTOs.usuario
 {
     public class UsuarioResponseDto
     {
-        public int IdUsuario { get; set; }
+        public long IdUsuario { get; set; }
         
         public string Nome { get; set; } = null!;
